+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # CHANGELOG
 
-## 0.5.4dev
+## 0.5.4 (2023-02-06)
 
-* [Fix] Adds community link to `ValueError` and `TypeError`
 * [Feature] Adds `%jupysql`/`%%jupysql` as alias for `%sql`/`%%sql`
+* [Fix] Adds community link to `ValueError` and `TypeError`
 
 ## 0.5.3 (2023-01-31)